Apply file-based routing to src/routes/user/teams/
