+++ src/routes/@shared/team-invites-schema.ts
@@ -1,6 +1,6 @@
 import { z } from 'zod'
 
-import type { ValidatedParamCtx, ValidatedParamJsonCtx } from './handler'
+import type { ValidatedParamJsonCtx } from './handler'
 
 import { dataRules } from './data-rules'
 import { permissionsSchema } from './permissions-schema'
@@ -21,19 +21,3 @@ export const inviteMemberBodySchema = z.object({
 export type TeamParams = z.infer<typeof teamParamsSchema>
 export type InviteMemberBody = z.infer<typeof inviteMemberBodySchema>
 export type InviteMemberCtx = ValidatedParamJsonCtx<TeamParams, InviteMemberBody>
-
-export const resendMemberInviteParamsSchema = z.object({
-  teamId: dataRules.id,
-  memberId: dataRules.id,
-})
-
-export type ResendMemberInviteParams = z.infer<typeof resendMemberInviteParamsSchema>
-export type ResendMemberInviteCtx = ValidatedParamCtx<ResendMemberInviteParams>
-
-export const cancelMemberInviteParamsSchema = z.object({
-  teamId: dataRules.id,
-  memberId: dataRules.id,
-})
-
-export type CancelMemberInviteParams = z.infer<typeof cancelMemberInviteParamsSchema>
-export type CancelMemberInviteCtx = ValidatedParamCtx<CancelMemberInviteParams>
