+++ src/routes/user/index.ts
@@ -1,2 +1,2 @@
-export { default as meRoute } from './me'
-export { default as teamsRoute } from './teams'
+export { meRoute } from './me'
+export { teamsRoute } from './teams'
