+++ src/routes/user/teams/$id/members/index.ts
@@ -1,0 +1,39 @@
+import { Hono } from 'hono'
+
+import type { AppContext } from '@/context'
+
+import { requestValidator, teamAccessMiddleware } from '@/middleware'
+
+import { teamsMemberByIdRoute } from './$id'
+import {
+  createMemberHandler,
+  getMemberDefaultPermissionsHandler,
+  getTeamMembersHandler,
+} from './handler'
+import { inviteMemberBodySchema, teamIdParamsSchema } from './schema'
+
+export const teamsMembersRoute = new Hono<AppContext>()
+
+teamsMembersRoute.get(
+  '/',
+  requestValidator('param', teamIdParamsSchema),
+  teamAccessMiddleware,
+  getTeamMembersHandler,
+)
+
+teamsMembersRoute.post(
+  '/',
+  requestValidator('param', teamIdParamsSchema),
+  requestValidator('json', inviteMemberBodySchema),
+  teamAccessMiddleware,
+  createMemberHandler,
+)
+
+teamsMembersRoute.get(
+  '/permissions',
+  requestValidator('param', teamIdParamsSchema),
+  teamAccessMiddleware,
+  getMemberDefaultPermissionsHandler,
+)
+
+teamsMembersRoute.route('/:memberId', teamsMemberByIdRoute)
