+++ src/routes/user/teams/$id/schema.ts
@@ -1,0 +1,7 @@
+export {
+  type TeamParamsCtx as TeamIdCtx,
+  type TeamParams as TeamIdParams,
+  teamParamsSchema as teamIdParamsSchema,
+  updateTeamBodySchema,
+  type UpdateTeamCtx,
+} from '@/routes/@shared'
