+++ src/routes/user/teams/index.ts
@@ -2,29 +2,8 @@ import { Hono } from 'hono'
 
 import type { AppContext } from '@/context'
 
-import { requestValidator, teamAccessMiddleware } from '@/middleware'
+import { teamByIdRoute } from './$id'
 
-import { getTeamHandler, updateTeamHandler } from './handler'
-import teamsMembersRoute from './members'
-import { teamParamsSchema, updateTeamBodySchema } from './schema'
+export const teamsRoute = new Hono<AppContext>()
 
-const teamsRoute = new Hono<AppContext>()
-
-teamsRoute.get(
-  '/teams/:teamId',
-  requestValidator('param', teamParamsSchema),
-  teamAccessMiddleware,
-  getTeamHandler,
-)
-
-teamsRoute.patch(
-  '/teams/:teamId',
-  requestValidator('param', teamParamsSchema),
-  requestValidator('json', updateTeamBodySchema),
-  teamAccessMiddleware,
-  updateTeamHandler,
-)
-
-teamsRoute.route('/teams/:teamId/members', teamsMembersRoute)
-
-export default teamsRoute
+teamsRoute.route('/teams/:teamId', teamByIdRoute)
