--- src/routes/user/teams/schema.ts
@@ -1,6 +1,0 @@
-export {
-  type TeamParamsCtx,
-  teamParamsSchema,
-  updateTeamBodySchema,
-  type UpdateTeamCtx,
-} from '../../@shared'
